# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpendBoundary — Policy-Gated Agentic Commerce Gateway",
    "description": "Merchant-side financial firewall for autonomous AI agents. Every cart is re-priced server-side in integer paise and evaluated against a deterministic 3-zone policy: ALLOW (autonomous zero-OTP debit), REVIEW (hosted payment link, human authorizes) or DENY (blocked at the gateway).",
    "version": "2.1.0"
  },
  "servers": [{ "url": "http://localhost:3000", "description": "Local SpendBoundary gateway" }],
  "paths": {
    "/api/catalogue": {
      "get": {
        "operationId": "searchCatalogue",
        "summary": "Search the merchant catalogue",
        "description": "Returns live stock and the authoritative integer paise price for each SKU. These are the only prices the gateway will honour.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Free text search over name, description, category and SKU."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "products": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Product" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "sku": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "pricePaise": { "type": "integer" },
          "priceFormatted": { "type": "string" },
          "stock": { "type": "integer" },
          "categoryWhitelisted": { "type": "boolean" }
        }
      },
...
    }
  }
}